arch2/services/upload/requirements.txt
flask
werkzeug
grpcio==1.75.1
grpcio-tools==1.75.1
protobuf==6.32.1
PyJWT
requests